+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: lugibone <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/04 12:54:35 by lugibone          #+#    #+#             */
-/*   Updated: 2019/04/05 17:02:19 by lugibone         ###   ########.fr       */
+/*   Updated: 2019/04/05 17:35:04 by lugibone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 
 	if (!s || !f)
 		return (NULL);
-	result = ft_strnew(ft_strlen((char*)s));
+	result = ft_strnew((int)ft_strlen((char*)s));
 	if (result == NULL)
 		return (NULL);
 	i = 0;
