+++ ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: lugibone <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/04 15:37:13 by lugibone          #+#    #+#             */
-/*   Updated: 2019/04/04 17:16:21 by lugibone         ###   ########.fr       */
+/*   Updated: 2019/04/05 18:01:25 by lugibone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,14 +28,14 @@ char	*ft_strtrim(char const *s)
 		len--;
 	if (len <= 0)
 		len = 0;
-	if (len == ft_strlen((char*)s))
+	if (len == (int)ft_strlen((char*)s))
 		return (ft_strdup((char*)s));
 	str = ft_strnew(len);
 	if (str == NULL)
 		return (NULL);
 	s += i;
 	i = -1;
-	while (++i < len)
+	while (++i < (int)len)
 		str[i] = *s++;
 	str[i] = '\0';
 	return (str);
